Apply STANDALONE configuration to all other gbserver CLI commands.
